ai_agent/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
huggingface-hub
sentence-transformers
scipy
scikit-learn
numpy
sounddevice
elevenlabs
PyPDF2
uuid
tqdm
python-multipart
langchain-community
pypdf
langchain-huggingface
httpx
httpx[http2]
passlib[bcrypt]
prometheus-fastapi-instrumentator
docker==5.0.3
prometheus_client
deepgram-sdk
faiss-cpu
pymongo
pydantic
euriai
pyaudio
jiwer 
pandas
pydub
asyncio
deepgram-sdk
pymupdf
pydantic[email]
fastapi[all]
pyjwt
python-jose[cryptography]
pycryptodome
wheel
pyppeteer 
openai==1.109.0
langchain==0.3.27
langchain-openai==0.3.33
langgraph==0.6.7
typing_extensions==4.14.1
pytz==2025.2

#python -m venv venv
#source venv/bin/activate  #mac
#venv\Scripts\activate #windows
#pip install -r requirements.txt
# uvicorn main:app --reload --port 8001
#http://localhost:8001/metrics
#for complete uninstall
#pip uninstall -r requirements.txt --yes
 
#pip install numpy==1.26.4

#uvicorn main:app --host 0.0.0.0 --port 8000 --reload
